--- NICE.Identity.Management/ClientApp/src/views/ServicesList/__tests__/webites.json
@@ -1,47 +1,0 @@
-[
-	{
-		"id": 1,
-		"serviceId": 1,
-		"host": "dev-eppi.nice.org.uk",
-		"environment": {
-		  "environmentId": 71,
-		  "name": "dev"
-		}
-	  },
-	  {
-		"id": 2,
-		"serviceId": 1,
-		"host": "test-eppi.nice.org.uk",
-		"environment": {
-		  "environmentId": 72,
-		  "name": "test"
-		}
-	  },
-	  {
-		"id": 3,
-		"serviceId": 2,
-		"host": "dev-comments.nice.org.uk",
-		"environment": {
-		  "environmentId": 71,
-		  "name": "dev"
-		}
-	  },
-	  {
-		"id": 4,
-		"serviceId": 2,
-		"host": "alpha-comments.nice.org.uk",
-		"environment": {
-		  "environmentId": 73,
-		  "name": "alpha"
-		}
-	  },
-	  {
-		"id": 5,
-		"serviceId": 3,
-		"host": "idam",
-		"environment": {
-		  "environmentId": 72,
-		  "name": "test"
-		}
-	  }
-]
