+++ NICE.Identity.Management/ClientApp/src/views/ServicesList/__tests__/websites.json
@@ -1,0 +1,67 @@
+[
+	{
+	  "id": 1,
+	  "serviceId": 1,
+	  "host": "dev-eppi.nice.org.uk",
+	  "environment": {
+		"environmentId": 71,
+		"name": "dev"
+	  },
+	  "service": {
+		"id": 1,
+		"name": "EPPI Reviewer"
+	  }
+	},
+	{
+	  "id": 2,
+	  "serviceId": 1,
+	  "host": "test-eppi.nice.org.uk",
+	  "environment": {
+		"environmentId": 72,
+		"name": "test"
+	  },
+	  "service": {
+		"id": 1,
+		"name": "EPPI Reviewer"
+	  }
+	},
+	{
+	  "id": 3,
+	  "serviceId": 2,
+	  "host": "dev-comments.nice.org.uk",
+	  "environment": {
+		"environmentId": 71,
+		"name": "dev"
+	  },
+	  "service": {
+		"id": 2,
+		"name": "Comment Collection"
+	  }
+	},
+	{
+	  "id": 4,
+	  "serviceId": 2,
+	  "host": "alpha-comments.nice.org.uk",
+	  "environment": {
+		"environmentId": 73,
+		"name": "alpha"
+	  },
+	  "service": {
+		"id": 2,
+		"name": "Comment Collection"
+	  }
+	},
+	{
+	  "id": 5,
+	  "serviceId": 3,
+	  "host": "idam",
+	  "environment": {
+		"environmentId": 72,
+		"name": "test"
+	  },
+	  "service": {
+		"id": 3,
+		"name": "Idam Docker"
+	  }
+	}
+  ]
